[Bronze V] Title: A/B, Time: 120 ms, Memory: 14348 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.java
@@ -6,9 +6,9 @@ public class Main{
     public static void main(String[]args)throws Exception{
         BufferedReader bf=new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st=new StringTokenizer(bf.readLine());
-        int a=Integer.parseInt(st.nextToken());
-        int b=Integer.parseInt(st.nextToken());
-        System.out.println((double)a/b);
+        double a=Integer.parseInt(st.nextToken());
+        double b=Integer.parseInt(st.nextToken());
+        System.out.println(a/b);
         
     }
 }
